+++ pyannote/database/util.py
@@ -26,6 +26,113 @@
 # AUTHORS
 # Hervé BREDIN - http://herve.niderb.fr
 
+import yaml
+from glob import glob
+
 
 class PyannoteDatabaseException(Exception):
     pass
+
+
+class FileFinder(object):
+    """Database file finder
+
+    Parameters
+    ----------
+    config_yml : str, optional
+        Path to database configuration file in YAML format.
+        See "Configuration file" sections for examples.
+        Defaults to '~/.pyannote/db.yml'.
+
+    Configuration file
+    ------------------
+    Here are a few examples of what is expected in the configuration file.
+
+    # all files are in the same directory
+    /path/to/files/{uri}.wav
+
+    # support for {database} placeholder
+    /path/to/{database}/files/{uri}.wav
+
+    # support for multiple databases
+    database1: /path/to/files/{uri}.wav
+    database2: /path/to/other/files/{uri}.wav
+
+    # files are spread over multiple directory
+    database3:
+      - /path/to/files/1/{uri}.wav
+      - /path/to/files/2/{uri}.wav
+
+    # supports * globbing
+    database4: /path/to/files/*/{uri}.wav
+
+    See also
+    --------
+    glob
+    """
+
+    def __init__(self, config_yml=None):
+        super(FileFinder, self).__init__()
+
+        if config_yml is None:
+            config_yml = '~/.pyannote/db.yml'
+        config_yml = os.path.expanduser(config_yml)
+
+        with open(config_yml, 'r') as fp:
+            self.config = yaml.load(fp)
+
+    def _find(self, config, uri=None, database=None, **kwargs):
+
+        found = []
+
+        # list of path templates
+        if isinstance(config, list):
+
+            for path_template in config:
+                path = path_template.format(uri=uri, database=database,
+                                            **kwargs)
+                found_ = glob(path)
+                found.extend(found_)
+
+        # database-indexed dictionary
+        elif isinstance(config, dict):
+
+            # if database identifier is not provided
+            # or does not exist in configuration file
+            # look into all databases...
+            if database is None or database not in config:
+                databases = list(config)
+            # if database identifier is provided AND exists
+            # only look into this very database
+            else:
+                databases = [database]
+
+            # iteratively look into selected databases
+            for database in databases:
+                found_ = self._find(config[database], uri=uri,
+                                    database=database, **kwargs)
+                found.extend(found_)
+
+        else:
+            path_template = config
+            path = path_template.format(uri=uri, database=database, **kwargs)
+            found_ = glob(path)
+            found.extend(found_)
+
+        return found
+
+    def __call__(self, uri=None, database=None, **kwargs):
+
+        # look for medium based on its uri and the database it belongs to
+        found = self._find(self.config, uri=uri, database=database, **kwargs)
+
+        if len(found) == 1:
+            return found[0]
+
+        elif len(found) == 0:
+            msg = 'Could not find file "{uri}".'
+            raise ValueError(msg.format(uri=uri))
+
+        else:
+            msg = 'Found {n} matches for file "{uri}"'
+            raise ValueError(msg.format(uri=uri, n=len(found)))